fix: token

+++ src/modules/cards/useCases/deleteCard/DeleteCardUseCase.ts
@@ -2,6 +2,7 @@ import { inject, injectable } from 'tsyringe';
 
 import { Card } from '@modules/cards/infra/typeorm/entities/Card';
 import { ICardsRepository } from '@modules/cards/repositories/ICardsRepository';
+import { AppError } from '@shared/errors/AppError';
 
 @injectable()
 class DeleteCardUseCase {
@@ -11,6 +12,8 @@ class DeleteCardUseCase {
   ) {}
 
   async execute(id: string): Promise<Card[]> {
+    const card = await this.cardsRepository.findById(id);
+    if (!card) throw new AppError('Card not found', 404);
     await this.cardsRepository.deleteById(id);
     return this.cardsRepository.list();
   }
